% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examen_pdf.R
\name{examen_pdf}
\alias{examen_pdf}
\alias{examen_pdf_solution}
\title{Convert to an examen PDF document}
\usage{
examen_pdf(solution = FALSE, suffix = "_question", id = FALSE,
  mcq = "oneparchoices", includes = NULL, pandoc_args = NULL, ...)

examen_pdf_solution(suffix = "_solution", ...)
}
\arguments{
\item{solution}{Turn ON or OFF the rendering of solution chunks (default is \code{FALSE})}

\item{suffix}{Suffix which is added to the filename (default is '_question' for 'unilur::examen_pdf' and '_solution' for 'unilur::examen_pdf_solution')}

\item{id}{Draw a student identification box}

\item{mcq}{Theme for the multiple choice questions (\code{oneparchoices}, \code{oneparchoicesalt}, \code{oneparcheckboxesalt} or \code{oneparcheckboxes})}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{Arguments passed to \code{pdf_document()}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
Format for converting from R Markdown to an examen PDF document.
}
\details{
See the inherited `rmarkdown::pdf_document` help page for additional arguments.
}
