% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_pdf.R
\name{to_latex_rgb}
\alias{to_latex_rgb}
\title{Helper function to construct the colour definition in latex (xcolor)}
\usage{
to_latex_rgb(x)
}
\arguments{
\item{x}{a vector containing the rgb colour values}
}
\description{
Builds up the string `\{RGB\}\{r, g, b\}` which needs to be completed: `\\definecolor\{name\}\{RGB\}\{r, g, b\}`
}
