% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_html.R
\name{tutorial_html}
\alias{tutorial_html}
\alias{tutorial_html_solution}
\title{Convert to a tutorial HTML document}
\usage{
tutorial_html(solution = FALSE, solution_suffix = "_solution",
  question_suffix = "_question", credit = FALSE, includes = NULL,
  css = NULL, extra_dependencies = NULL, ...)

tutorial_html_solution(...)
}
\arguments{
\item{solution}{Turn ON or OFF the rendering of solution chunks (default is \code{FALSE})}

\item{solution_suffix}{Suffix which is added to the filename when \code{solution = TRUE} (default is '_solution')}

\item{question_suffix}{Suffix which is added to the filename when \code{solution = FALSE} (default is '_question')}

\item{credit}{Turn ON or OFF the footer showing a link to the unilur homepage (default is \code{FALSE})}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{css}{One or more css files to include}

\item{extra_dependencies}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}

\item{...}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to a tutorial HTML document.
}
\details{
See the inherited `rmarkdown::html_document` help page for additional arguments.
}
