% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_pdf.R
\name{tutorial_pdf}
\alias{tutorial_pdf}
\alias{tutorial_pdf_solution}
\title{Convert to a tutorial PDF document}
\usage{
tutorial_pdf(solution = FALSE, suffix = "_question",
  pandoc_args = NULL, ...)

tutorial_pdf_solution(suffix = "_solution", ...)
}
\arguments{
\item{solution}{Turn ON or OFF the rendering of solution chunks (default is \code{FALSE})}

\item{suffix}{Suffix which is added to the filename (default is '_question' for 'unilur::tutorial_pdf' and '_solution' for 'unilur::tutorial_pdf_solution')}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{Arguments passed to \code{pdf_document()}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
Format for converting from R Markdown to a tutorial PDF document.
}
\details{
See the inherited `rmarkdown::pdf_document` help page for additional arguments.
}
