% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unilur_base.R
\name{unilur_base}
\alias{unilur_base}
\title{Define common settings for pdf and HTML output format}
\usage{
unilur_base(format, is_solution_format, suffix)
}
\arguments{
\item{format}{a format to adapt (rmarkdown::html_document or rmarkdown::pdf_document)}

\item{is_solution_format}{`logical` should the output format render solution chunks or not}

\item{suffix}{Suffix which is added to the filename (to generate two separate files from the same source without overwriting)}
}
\value{
R Markdown output format
}
\description{
Set the post_processor, pre_knit as well as the solution theme in opts_template (which are common to all unilur formats)
}
