% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_rmd.R
\name{answer_rmd}
\alias{answer_rmd}
\title{Convert to a tutorial answer Rmarkdown file}
\usage{
answer_rmd(yaml = NULL, suffix = "_answer", exclude = NULL)
}
\arguments{
\item{yaml}{a list to override the \code{title}, \code{author} and \code{output} of the answer Rmd file.}

\item{suffix}{Suffix which is added to the filename (default is '_answer')}

\item{exclude_chunk}{keyword to use for excluding specific chunks in answer templates (default NULL)}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
Format to generate a Rmd file with solutions chunks being replaced by empty answer chunks.
}
