% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_html.R
\name{tutorial_html}
\alias{tutorial_html}
\alias{tutorial_html_solution}
\title{Convert to a tutorial HTML document}
\usage{
tutorial_html(
  solution = FALSE,
  suffix = "_question",
  includes = NULL,
  css = NULL,
  extra_dependencies = NULL,
  ...
)

tutorial_html_solution(suffix = "_solution", ...)
}
\arguments{
\item{solution}{Turn ON or OFF the rendering of solution chunks (default is \code{FALSE})}

\item{suffix}{Suffix which is added to the filename (default is '_question' for 'unilur::tutorial_html' and '_solution' for 'unilur::tutorial_html_solution')}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{css}{One or more css files to include}

\item{extra_dependencies}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link[rmarkdown]{html_document_base}}}

\item{...}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link[rmarkdown]{html_document_base}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
Format for converting from R Markdown to a tutorial HTML document.
}
\details{
See the inherited `rmarkdown::html_document` help page for additional arguments.
}
