% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_pdf.R
\name{tutorial_pdf}
\alias{tutorial_pdf}
\alias{tutorial_pdf_solution}
\title{Convert to a tutorial PDF document}
\usage{
tutorial_pdf(...)

tutorial_pdf_solution(...)
}
\arguments{
\item{...}{Arguments passed to \code{pdf_document()}.}

\item{solution}{Turn ON or OFF the rendering of solution chunks (default is \code{FALSE})}

\item{solution_suffix}{Suffix which is added to the filename when \code{solution = TRUE} (default is '_solution')}

\item{question_suffix}{Suffix which is added to the filename when \code{solution = FALSE} (default is '_question')}

\item{credit}{Turn ON or OFF the footer showing a link to the unilur homepage (default is \code{FALSE})}
}
\value{
R Markdown output format to pass to \code{\link{render}}
}
\description{
Format for converting from R Markdown to a tutorial PDF document.
}
\details{
See the inherited `rmarkdown::pdf_document` help page for additional arguments.
}
